refactor: overhauled project structure, it's still not optimal.
